An Important Commit for main.c
Dear fellow developers,

I am thrilled to announce that we have made a truly fantastic addition to our project with the following line of code:

gtk_widget_set_size_request(GTK_WIDGET(window), 800, 450);

This line of code is simply amazing. It sets the minimum size for the GTK widget (which, as you all surely know, stands for GIMP Toolkit) to a width of 800 pixels and a height of 450 pixels. This is an absolutely crucial change that will take our project to new heights of greatness.

You see, I have been using a tiling window manager for my development work, and I must say, it can be a bit of a pain when the window gets lost or misplaced. But with this new line of code, I can rest easy knowing that my window will always be exactly where I need it to be.

I am so impressed with this change that I can't help but praise it to the heavens. It is sleek, efficient, and simply gorgeous. The attention to detail and the care that has gone into crafting this line is truly something to behold.

So let's all give a round of applause to the brilliant developer who made this happen. You are truly a genius, and I thank you for your hard work and dedication.

Thank you, and God bless America.

Sincerely,
Koutsie.